+++ backend/service_starter.go
@@ -2,6 +2,7 @@ package backend
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -9,7 +10,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	vgclose "code.vegaprotocol.io/vega/libs/close"
 	vgjob "code.vegaprotocol.io/vega/libs/job"
 	vgzap "code.vegaprotocol.io/vega/libs/zap"
 	"code.vegaprotocol.io/vega/paths"
@@ -17,7 +17,6 @@ import (
 	"code.vegaprotocol.io/vega/wallet/api/interactor"
 	"code.vegaprotocol.io/vega/wallet/service"
 	svcStoreV1 "code.vegaprotocol.io/vega/wallet/service/store/v1"
-	serviceV1 "code.vegaprotocol.io/vega/wallet/service/v1"
 	"code.vegaprotocol.io/vega/wallet/service/v2/connections"
 	"code.vegaprotocol.io/vegawallet-desktop/app"
 	"code.vegaprotocol.io/vegawallet-desktop/os/notification"
@@ -25,10 +24,10 @@ import (
 	"go.uber.org/zap"
 )
 
+var ErrNoChanResponseForTraceID = errors.New("no response expected for that traceID")
+
 type ServiceStarter struct {
 	serviceStarter *service.Starter
-
-	closer *vgclose.Closer
 
 	url string
 
@@ -36,14 +35,15 @@ type ServiceStarter struct {
 
 	logFilePath string
 
-	receptionChan chan interactor.Interaction
-	responseChan  chan interactor.Interaction
+	receptionChan <-chan interactor.Interaction
 
 	isRunning atomic.Bool
 
 	mu        sync.Mutex
 	log       *zap.Logger
 	jobRunner *vgjob.Runner
+
+	responseChansByTraceID map[string]chan<- interactor.Interaction
 }
 
 func (s *ServiceStarter) StartService(ctx context.Context, network string) (err error) {
@@ -64,17 +64,17 @@ func (s *ServiceStarter) StartService(ctx context.Context, network string) (err 
 		}
 	}()
 
-	svcURL, errChan, err := s.serviceStarter.Start(s.jobRunner, network, true)
+	rc, err := s.serviceStarter.Start(s.jobRunner, network, true)
 	if err != nil {
 		s.log.Error("Failed to start HTTP server", zap.Error(err))
 		return err
 	}
-	s.url = svcURL
+	s.url = rc.ServiceURL
 
-	s.log.Info("Starting HTTP service", zap.String("url", svcURL))
+	s.log.Info("Starting HTTP service", zap.String("url", rc.ServiceURL))
 
 	s.jobRunner.Go(func(jobCtx context.Context) {
-		s.listenToServiceRuntimeError(jobCtx, errChan)
+		s.listenToServiceRuntimeError(jobCtx, rc.ErrCh)
 	})
 
 	s.jobRunner.Go(func(jobCtx context.Context) {
@@ -84,7 +84,7 @@ func (s *ServiceStarter) StartService(ctx context.Context, network string) (err 
 	// We warn the front-end by sending events when the service is unhealthy.
 	// This is done so the front-end doesn't have to do it.
 	s.jobRunner.Go(func(jobCtx context.Context) {
-		s.monitorServiceHealth(jobCtx, svcURL)
+		s.monitorServiceHealth(jobCtx, rc.ServiceURL)
 	})
 
 	return nil
@@ -104,7 +104,6 @@ func (s *ServiceStarter) Close() {
 	if s.isRunning.Load() {
 		s.jobRunner.StopAllJobs()
 	}
-	s.closer.CloseAll()
 	s.reset()
 }
 
@@ -137,6 +136,26 @@ func (s *ServiceStarter) ServiceHealth() HealthCheckStatus {
 func (s *ServiceStarter) listenToIncomingInteractions(ctx context.Context) {
 	log := s.log.Named("service-interactions-listener")
 	log.Info("Listening to incoming interactions")
+
+	// Since the reception channel is not re-created on every service restart,
+	// we need to purge it to avoid leaving interaction a the previous session
+	// be treated as from the newer session.
+	defer func() {
+		log.Info("Purging interactions left-over")
+		defer log.Info("Left-over interactions purged")
+		for {
+			select {
+			case interaction, ok := <-s.receptionChan:
+				if !ok {
+					return
+				}
+				log.Debug("Purged interaction", zap.String("name", string(interaction.Name)))
+			default:
+				return
+			}
+		}
+	}()
+
 	for {
 		select {
 		case <-ctx.Done():
@@ -169,31 +188,60 @@ func (s *ServiceStarter) emitReceivedInteraction(ctx context.Context, log *zap.L
 		runtime.WindowShow(ctx)
 	}
 
+	switch data := interaction.Data.(type) {
+	case interactor.RequestTransactionReviewForSigning:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestTransactionReviewForSending:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestPermissionsReview:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestPassphrase:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestWalletSelection:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestWalletConnectionReview:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	case interactor.RequestTransactionReviewForChecking:
+		s.responseChansByTraceID[interaction.TraceID] = data.ResponseCh
+	}
+
 	runtime.EventsEmit(ctx, NewInteractionEvent, interaction)
+}
+
+func (s *ServiceStarter) RespondToInteraction(interaction interactor.Interaction) error {
+	ch, ok := s.responseChansByTraceID[interaction.TraceID]
+	if !ok {
+		return ErrNoChanResponseForTraceID
+	}
+
+	ch <- interaction
+
+	delete(s.responseChansByTraceID, interaction.TraceID)
+
+	return nil
 }
 
 func (s *ServiceStarter) listenToServiceRuntimeError(jobCtx context.Context, errChan <-chan error) {
 	log := s.log.Named("service-runtime-error-listener")
 	log.Info("Listening to service runtime error")
 
-	for {
-		select {
-		case <-jobCtx.Done():
-			log.Info("Stopping the listening to the service runtime error")
+	select {
+	case <-jobCtx.Done():
+		break
+	case err, ok := <-errChan:
+		if !ok {
 			return
-		case err, ok := <-errChan:
-			if !ok {
-				return
-			}
-			s.log.Error("An error occurred while running the service", zap.Error(err))
-			s.StopService()
-			runtime.EventsEmit(jobCtx, ServiceStoppedWithError, struct {
-				Error error
-			}{
-				Error: err,
-			})
 		}
+		s.log.Error("An error occurred while running the service", zap.Error(err))
+		s.StopService()
+		runtime.EventsEmit(jobCtx, ServiceStoppedWithError, struct {
+			Error error
+		}{
+			Error: err,
+		})
 	}
+
+	log.Info("Stopping the listening to the service runtime error")
 }
 
 func (s *ServiceStarter) monitorServiceHealth(ctx context.Context, svcURL string) {
@@ -237,7 +285,7 @@ func (s *ServiceStarter) queryServiceHealth(ctx context.Context, log *zap.Logger
 		log.Debug("The service is healthy")
 		s.latestHealthState = ServiceIsHealthy
 	} else {
-		log.Warn("The service is reachable but is not healthy", zap.String("code", response.Status))
+		log.Warn("The service is reachable but is unhealthy", zap.String("code", response.Status))
 		s.latestHealthState = ServiceIsUnhealthy
 	}
 	runtime.EventsEmit(ctx, string(s.latestHealthState))
@@ -247,17 +295,17 @@ func (s *ServiceStarter) reset() {
 	s.url = ""
 	s.logFilePath = ""
 	s.latestHealthState = UnknownStatus
-	s.closer = vgclose.NewCloser()
 	s.receptionChan = nil
-	s.responseChan = nil
 	s.jobRunner = nil
 	s.isRunning.Store(false)
 }
 
-func NewServiceStarter(vegaPaths paths.Paths, log *zap.Logger, svcStore *svcStoreV1.Store, walletStore walletapi.WalletStore, netStore walletapi.NetworkStore, connectionsManager *connections.Manager) (*ServiceStarter, error) {
+func NewServiceStarter(vegaPaths paths.Paths, log *zap.Logger, svcStore *svcStoreV1.Store, walletStore walletapi.WalletStore, netStore walletapi.NetworkStore, connectionsManager *connections.Manager, apiInteractor walletapi.Interactor, receptionChan <-chan interactor.Interaction) (*ServiceStarter, error) {
 	s := &ServiceStarter{}
 	s.reset()
 	s.log = log
+	s.receptionChan = receptionChan
+	s.responseChansByTraceID = map[string]chan<- interactor.Interaction{}
 
 	loggerBuilderFunc := func(levelName string) (*zap.Logger, zap.AtomicLevel, error) {
 		svcLog, svcLogPath, level, err := buildServiceLogger(vegaPaths, paths.WalletServiceLogsHome, levelName)
@@ -268,31 +316,15 @@ func NewServiceStarter(vegaPaths paths.Paths, log *zap.Logger, svcStore *svcStor
 		return svcLog, level, nil
 	}
 
-	policyBuilderFunc := func(_ context.Context) serviceV1.Policy {
-		return &unsupportedV1APIPolicy{}
-	}
-
-	interactorBuilderFunc := func(ctx context.Context) walletapi.Interactor {
-		receptionChan := make(chan interactor.Interaction, 100)
-		responseChan := make(chan interactor.Interaction, 100)
-		s.receptionChan = receptionChan
-		s.responseChan = responseChan
-
-		s.closer.Add(func() {
-			close(receptionChan)
-			close(responseChan)
-		})
-
-		return interactor.NewSequentialInteractor(ctx, receptionChan, responseChan)
-	}
+	policy := &unsupportedV1APIPolicy{}
 
 	serviceStarter := service.NewStarter(
 		walletStore,
 		netStore,
 		svcStore,
 		connectionsManager,
-		policyBuilderFunc,
-		interactorBuilderFunc,
+		policy,
+		apiInteractor,
 		loggerBuilderFunc,
 	)